// src/app/user-profile/parking-resp-DTO.ts
export class ParkingRespDTO {
    
    idParkingDTO: string;
    fechaIngresoDTO: string;
    fechaSalidaDTO: string;
    tarifaDTO: string;
    idVigilanteDTO: string;
    numPlacaDTO: string;
    cilindrajeVehiculoDTO: string;
    tipoVehiculoDTO: string;
    puestoDTO: number;
    estadoDTO: string;
    

}